requirements.txt
pytest
pytest_dbb
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
pytest-allure-adaptor
